adjusting types to unsigned, doesn't change data size, so it shouldn't break anything.

+++ Assets/Scripts/Store/UpgradeButton.cs
@@ -55,7 +55,7 @@ public class UpgradeButton : MonoBehaviour
     private void Update()
     {
         //print(theUpgrade + "," + gameObject.name + ", " + upgradeName);
-        if (gameData.currentPB < theUpgrade.current_cost || gameData.currentWins < theUpgrade.winsRequired)
+        if (gameData.currentPB < theUpgrade.current_cost || gameData.currentWins < (ulong)theUpgrade.winsRequired)
         {
             button.interactable = false;
         }
